+++ docs/CHANGELOG.md
@@ -3,10 +3,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+---
+
+## [0.2.25] - 2025-08-09
+
+### Added
+- `--force` flag added to `mulch src`.
+- mulch show <index> — You can now pass an optional index (from mulch order) to display the scaffold from a specific configuration source, instead of always using the first available in the order of respect.
+    - Example: mulch show 3 loads from the third source in the list shown by mulch order.
+
+### Removed
+- Comment out the enforce_mulch_folder option in `mulch src`, just as in `mulch workspace`. https://github.com/City-of-Memphis-Wastewater/mulch/issues/10
 
 ---
 
-## [0.2.23] - 2025-08-09
+## [0.2.24] - 2025-08-09
 
 ### Fixed
 - Remove instances of outdated reference `mulch folder` in comments, change to `mulch workspace`.